requirements.txt
astroid==2.4.1
certifi==2021.5.30
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
paho-mqtt==1.5.1
pylint==2.5.2
toml==0.10.1
wincertstore==0.2
wrapt==1.12.1
